requirements.txt
numpy
pandas
python-dotenv # ye envirnment veriable set kiye h use read krne me help krega
mysql-connector-python
pymysql
scikit-learn
#-e .